refactored pages to remove metadata from client components
