+++ src/app/error.tsx
@@ -2,13 +2,7 @@
 import { PageContainer } from '@/components/utils/page-container';
 import Image from 'next/image';
 import oops from '../../public/static/images/pages/error/oops.png';
-import { HOME_PAGE_TITLE } from '@/constants/metadata';
-import type { Metadata } from 'next';
 import styles from './error.module.scss';
-
-export const metadata: Metadata = {
-  title: `${HOME_PAGE_TITLE} | Error`,
-};
 
 export default function Error() {
   return (
