+++ src/app/register/addresses/page.tsx
@@ -1,4 +1,10 @@
 import dynamic from 'next/dynamic';
+import { HOME_PAGE_TITLE } from '@/constants/metadata';
+import type { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: `${HOME_PAGE_TITLE} | Register to Vote`,
+};
 
 /*
   Render the addresses form on the client side to prevent hydration errors
